+++ src/appmixer/akamai/bundle.json
@@ -1,6 +1,6 @@
 {
   "name": "appmixer.akamai",
-  "version": "1.0.3",
+  "version": "1.1.0",
   "changelog": {
       "1.0.1": [
           "Initial version"
@@ -12,6 +12,9 @@
       "1.0.3": [
         "UpsertListEntries: updated tooltip for IPs as 255.255.255.255 would return an error.",
         "Lists select now show the type of the List."
+      ],
+      "1.1.0": [
+        "UpsertListEntries, DeleteListEntries: added support for entering IPs as JSON."
       ]
   }
 }
