+++ src/appmixer/akamai/clientlist/ActivateList/ActivateList.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { generateAuthorizationHeader } = require('../../signature');
+const { generateAuthorizationHeader } = require('../../lib');
 
 module.exports = {
     async receive(context) {
