+++ src/appmixer/akamai/clientlist/DeleteListEntries/component.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.0.1",
+  "version": "1.1.0",
   "name": "appmixer.akamai.clientlist.DeleteListEntries",
   "author": "Appmixer <info@appmixer.com>",
   "description": "Delete client list entries.",
@@ -42,7 +42,7 @@
                     "index": 1,
                     "type": "text",
                     "label": "IP or IPs",
-                    "tooltip": "Value of the entry, which is either an IP address, an Autonomous System Number (ASN), a Geolocation, a TLS fingerprint, or a file hash. Multiple values can be added at once but need to be separated by a comma. Example: <code>1.1.1.1,255.255.255.255</code>"
+                    "tooltip": "Value of the entry, which is either an IP address, an Autonomous System Number (ASN), a Geolocation, a TLS fingerprint, or a file hash. Multiple values can be added at once but need to be separated by a comma. Example: <code>1.1.1.1,255.255.255.255</code>. JSON format is also supported, for example: [\"1.1.1.1\", \"2.2.2.2\"]."
                 }
             }
         }
