+++ src/appmixer/akamai/clientlist/UpsertListEntries/component.json
@@ -1,5 +1,5 @@
 {
-  "version": "1.0.3",
+  "version": "1.1.0",
   "name": "appmixer.akamai.clientlist.UpsertListEntries",
   "author": "Appmixer <info@appmixer.com>",
   "description": "Append or update client list entries and activates the list.",
@@ -43,7 +43,7 @@
                     "type": "text",
                     "index": 1,
                     "label": "IP or IPs",
-                    "tooltip": "Value of the entry, which is either an IP address, an Autonomous System Number (ASN), a Geolocation, a TLS fingerprint, or a file hash. Multiple values can be added at once but need to be separated by a comma. It will then use the same Description and Expiration Date. Example: <code>1.1.1.1,192.168.1.17</code>"
+                    "tooltip": "Value of the entry, which is either an IP address, an Autonomous System Number (ASN), a Geolocation, a TLS fingerprint, or a file hash. Multiple values can be added at once but need to be separated by a comma. It will then use the same Description and Expiration Date. Example: <code>1.1.1.1,192.168.1.17</code>. JSON format is also supported, for example: [\"1.1.1.1\", \"2.2.2.2\"]."
                 },
                 "network": {
                     "type": "select",
